octoshop/requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
cairocffi==1.2.0
CairoSVG==2.5.2
celery==5.1.2
cffi==1.14.6
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cssselect2==0.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.7
django-cleanup==5.2.0
django-debug-toolbar==3.2.2
django-filter==21.1
django-htmx==1.2.1
django-import-export==2.6.0
django-js-asset==1.2.2
django-mptt==0.13.3
django-mptt-admin==2.1.0
django-nested-admin==3.3.3
django-tinymce==3.3.0
et-xmlfile==1.1.0
html5lib==1.1
kombu==5.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.3.2
prompt-toolkit==3.0.20
pycparser==2.20
pyphen==0.11.0
python-decouple==3.4
python-monkey-business==1.0.0
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.2
tablib==3.0.0
tinycss2==1.1.0
vine==5.0.0
wcwidth==0.2.5
WeasyPrint==51
webencodings==0.5.1
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
